requirements.txt
psycopg2
pandas
altair
plotly.express
streamlit
seaborn
matplotlib.pyplot